+++ l2geth/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.5.28
+
+### Patch Changes
+
+- ac0f14f5: Fix state dumper
+- 5005da9a: Fixes a small l2geth bug when trying to dump state
+
+## 0.5.27
+
+### Patch Changes
+
+- 596c974e: Kick the build
+
 ## 0.5.26
 
 ### Patch Changes